moved angular code into individual AMD modules
